requirements.txt
sklearn
pandas
numpy
scipy
mlflow
s3
boto3
fastapi
uvicorn[standart]
argparse
flake8
python-dotenv
